requirements.txt
toil[all]
django>=2.1,<2.2
celery>=4.2,<4.3
django-celery-results
djangorestframework
gitpython
ipython
django-debug-toolbar
uwsgi
cwl-utils
